+++ fsl_mrs/utils/fmrs_tools/__init__.py
@@ -1,2 +1,3 @@
 from .contrasts import create_contrasts, Contrast
 from .flame import flameo_wrapper
+from .scalings import fmrs_internal_reference